+++ warp/fem/quadrature/quadrature.py
@@ -1,12 +1,19 @@
-from typing import Any
+from typing import Any, Optional
 
 import warp as wp
-from warp.fem import cache, domain
+from warp.fem import cache
+from warp.fem.domain import GeometryDomain
 from warp.fem.geometry import Element
 from warp.fem.space import FunctionSpace
-from warp.fem.types import Coords, ElementIndex
+from warp.fem.types import NULL_ELEMENT_INDEX, Coords, ElementIndex, QuadraturePointIndex
 
 from ..polynomial import Polynomial
+
+
+@wp.struct
+class QuadraturePointElementIndex:
+    domain_element_index: ElementIndex
+    qp_index_in_element: int
 
 
 class Quadrature:
@@ -18,7 +25,7 @@ class Quadrature:
 
         pass
 
-    def __init__(self, domain: domain.GeometryDomain):
+    def __init__(self, domain: GeometryDomain):
         self._domain = domain
 
     @property
@@ -30,52 +37,197 @@ class Quadrature:
         """
         Value of the argument to be passed to device
         """
-        arg = RegularQuadrature.Arg()
+        arg = Quadrature.Arg()
         return arg
 
     def total_point_count(self):
-        """Total number of quadrature points over the domain"""
+        """Number of unique quadrature points that can be indexed by this rule.
+        Returns a number such that `point_index()` is always smaller than this number.
+        """
         raise NotImplementedError()
+
+    def evaluation_point_count(self):
+        """Number of quadrature points that needs to be evaluated, mostly for internal purposes.
+        If the indexing scheme is sparse, or if a quadrature point is shared among multiple elements
+        (e.g, nodal quadrature), `evaluation_point_count` may be different than `total_point_count()`.
+        Returns a number such that `evaluation_point_index()` is always smaller than this number.
+        """
+        return self.total_point_count()
 
     def max_points_per_element(self):
         """Maximum number of points per element if known, or ``None`` otherwise"""
         return None
 
     @staticmethod
-    def point_count(elt_arg: "domain.GeometryDomain.ElementArg", qp_arg: Arg, element_index: ElementIndex):
+    def point_count(
+        elt_arg: "GeometryDomain.ElementArg",
+        qp_arg: Arg,
+        domain_element_index: ElementIndex,
+        geo_element_index: ElementIndex,
+    ):
         """Number of quadrature points for a given element"""
         raise NotImplementedError()
 
     @staticmethod
     def point_coords(
-        elt_arg: "domain.GeometryDomain.ElementArg", qp_arg: Arg, element_index: ElementIndex, qp_index: int
+        elt_arg: "GeometryDomain.ElementArg",
+        qp_arg: Arg,
+        domain_element_index: ElementIndex,
+        geo_element_index: ElementIndex,
+        element_qp_index: int,
     ):
         """Coordinates in element of the element's qp_index'th quadrature point"""
         raise NotImplementedError()
 
     @staticmethod
     def point_weight(
-        elt_arg: "domain.GeometryDomain.ElementArg", qp_arg: Arg, element_index: ElementIndex, qp_index: int
+        elt_arg: "GeometryDomain.ElementArg",
+        qp_arg: Arg,
+        domain_element_index: ElementIndex,
+        geo_element_index: ElementIndex,
+        element_qp_index: int,
     ):
         """Weight of the element's qp_index'th quadrature point"""
         raise NotImplementedError()
 
     @staticmethod
     def point_index(
-        elt_arg: "domain.GeometryDomain.ElementArg",
+        elt_arg: "GeometryDomain.ElementArg",
         qp_arg: Arg,
         domain_element_index: ElementIndex,
         geo_element_index: ElementIndex,
         element_qp_index: int,
     ):
-        """Global index of the element's qp_index'th  quadrature point"""
+        """
+        Global index of the element's qp_index'th  quadrature point.
+        May be shared among elements.
+        This is what determines `qp_index` in integrands' `Sample` arguments.
+        """
+        raise NotImplementedError()
+
+    @staticmethod
+    def point_evaluation_index(
+        elt_arg: "GeometryDomain.ElementArg",
+        qp_arg: Arg,
+        domain_element_index: ElementIndex,
+        geo_element_index: ElementIndex,
+        element_qp_index: int,
+    ):
+        """Quadrature point index according to evaluation order.
+        Quadrature points for distinct elements must have different evaluation indices.
+        Mostly for internal/parallelization purposes.
+        """
         raise NotImplementedError()
 
     def __str__(self) -> str:
         return self.name
 
+    # By default cache the mapping from evaluation point indices to domain elements
 
-class RegularQuadrature(Quadrature):
+    ElementIndexArg = wp.array(dtype=QuadraturePointElementIndex)
+
+    @cache.cached_arg_value
+    def element_index_arg_value(self, device):
+        """Builds a map from quadrature point evaluation indices to their index in the element to which they belong"""
+
+        @cache.dynamic_kernel(f"{self.name}{self.domain.name}")
+        def quadrature_point_element_indices(
+            qp_arg: self.Arg,
+            domain_arg: self.domain.ElementArg,
+            domain_index_arg: self.domain.ElementIndexArg,
+            result: wp.array(dtype=QuadraturePointElementIndex),
+        ):
+            domain_element_index = wp.tid()
+            element_index = self.domain.element_index(domain_index_arg, domain_element_index)
+
+            qp_point_count = self.point_count(domain_arg, qp_arg, domain_element_index, element_index)
+            for k in range(qp_point_count):
+                qp_eval_index = self.point_evaluation_index(domain_arg, qp_arg, domain_element_index, element_index, k)
+                result[qp_eval_index] = QuadraturePointElementIndex(domain_element_index, k)
+
+        null_qp_index = QuadraturePointElementIndex()
+        null_qp_index.domain_element_index = NULL_ELEMENT_INDEX
+        result = wp.full(
+            value=null_qp_index,
+            shape=(self.evaluation_point_count()),
+            dtype=QuadraturePointElementIndex,
+            device=device,
+        )
+        wp.launch(
+            quadrature_point_element_indices,
+            device=result.device,
+            dim=self.domain.element_count(),
+            inputs=[
+                self.arg_value(result.device),
+                self.domain.element_arg_value(result.device),
+                self.domain.element_index_arg_value(result.device),
+                result,
+            ],
+        )
+
+        return result
+
+    @wp.func
+    def evaluation_point_element_index(
+        element_index_arg: wp.array(dtype=QuadraturePointElementIndex),
+        qp_eval_index: QuadraturePointIndex,
+    ):
+        """Maps from quadrature point evaluation indices to their index in the element to which they belong
+        If the quadrature poin does not exist, should return NULL_ELEMENT_INDEX as the domain element index
+        """
+
+        element_index = element_index_arg[qp_eval_index]
+        return element_index.domain_element_index, element_index.qp_index_in_element
+
+
+class _QuadratureWithRegularEvaluationPoints(Quadrature):
+    """Helper sublcass for quadrature formulas which use a uniform number of
+    evaluations points per element. Avoids building explicit mapping"""
+
+    def __init__(self, domain: GeometryDomain, N: int):
+        super().__init__(domain)
+        self._EVALUATION_POINTS_PER_ELEMENT = N
+
+        self.point_evaluation_index = self._make_regular_point_evaluation_index()
+        self.evaluation_point_element_index = self._make_regular_evaluation_point_element_index()
+
+    ElementIndexArg = Quadrature.Arg
+    element_index_arg_value = Quadrature.arg_value
+
+    def evaluation_point_count(self):
+        return self.domain.element_count() * self._EVALUATION_POINTS_PER_ELEMENT
+
+    def _make_regular_point_evaluation_index(self):
+        N = self._EVALUATION_POINTS_PER_ELEMENT
+
+        @cache.dynamic_func(suffix=f"{self.name}")
+        def evaluation_point_index(
+            elt_arg: self.domain.ElementArg,
+            qp_arg: self.Arg,
+            domain_element_index: ElementIndex,
+            element_index: ElementIndex,
+            qp_index: int,
+        ):
+            return N * domain_element_index + qp_index
+
+        return evaluation_point_index
+
+    def _make_regular_evaluation_point_element_index(self):
+        N = self._EVALUATION_POINTS_PER_ELEMENT
+
+        @cache.dynamic_func(suffix=f"{N}")
+        def quadrature_evaluation_point_element_index(
+            qp_arg: Quadrature.Arg,
+            qp_index: QuadraturePointIndex,
+        ):
+            domain_element_index = qp_index // N
+            index_in_element = qp_index - domain_element_index * N
+            return domain_element_index, index_in_element
+
+        return quadrature_evaluation_point_element_index
+
+
+class RegularQuadrature(_QuadratureWithRegularEvaluationPoints):
     """Regular quadrature formula, using a constant set of quadrature points per element"""
 
     @wp.struct
@@ -112,16 +264,15 @@ class RegularQuadrature(Quadrature):
 
     def __init__(
         self,
-        domain: domain.GeometryDomain,
+        domain: GeometryDomain,
         order: int,
         family: Polynomial = None,
     ):
-        super().__init__(domain)
-
+        self._formula = RegularQuadrature.CachedFormula.get(domain.reference_element(), order, family)
         self.family = family
         self.order = order
 
-        self._formula = RegularQuadrature.CachedFormula.get(domain.reference_element(), order, family)
+        super().__init__(domain, self._formula.count)
 
         self.point_count = self._make_point_count()
         self.point_index = self._make_point_index()
@@ -212,10 +363,10 @@ class NodalQuadrature(Quadrature):
     any assumption about orthogonality of shape functions, and is thus safe to use for arbitrary integrands.
     """
 
-    def __init__(self, domain: domain.GeometryDomain, space: FunctionSpace):
-        super().__init__(domain)
-
+    def __init__(self, domain: Optional[GeometryDomain], space: FunctionSpace):
         self._space = space
+
+        super().__init__(domain)
 
         self.Arg = self._make_arg()
 
@@ -223,6 +374,7 @@ class NodalQuadrature(Quadrature):
         self.point_index = self._make_point_index()
         self.point_coords = self._make_point_coords()
         self.point_weight = self._make_point_weight()
+        self.point_evaluation_index = self._make_point_evaluation_index()
 
     @property
     def name(self):
@@ -300,8 +452,26 @@ class NodalQuadrature(Quadrature):
 
         return point_index
 
+    def evaluation_point_count(self):
+        return self.domain.element_count() * self._space.topology.MAX_NODES_PER_ELEMENT
 
-class ExplicitQuadrature(Quadrature):
+    def _make_point_evaluation_index(self):
+        N = self._space.topology.MAX_NODES_PER_ELEMENT
+
+        @cache.dynamic_func(suffix=self.name)
+        def evaluation_point_index(
+            elt_arg: self.domain.ElementArg,
+            qp_arg: self.Arg,
+            domain_element_index: ElementIndex,
+            element_index: ElementIndex,
+            qp_index: int,
+        ):
+            return N * domain_element_index + qp_index
+
+        return evaluation_point_index
+
+
+class ExplicitQuadrature(_QuadratureWithRegularEvaluationPoints):
     """Quadrature using explicit per-cell points and weights.
 
     The number of quadrature points per cell is assumed to be constant and deduced from the shape of the points and weights arrays.
@@ -321,11 +491,7 @@ class ExplicitQuadrature(Quadrature):
         points: wp.array2d(dtype=Coords)
         weights: wp.array2d(dtype=float)
 
-    def __init__(
-        self, domain: domain.GeometryDomain, points: "wp.array2d(dtype=Coords)", weights: "wp.array2d(dtype=float)"
-    ):
-        super().__init__(domain)
-
+    def __init__(self, domain: GeometryDomain, points: "wp.array2d(dtype=Coords)", weights: "wp.array2d(dtype=float)"):
         if points.shape != weights.shape:
             raise ValueError("Points and weights arrays must have the same shape")
 
@@ -343,7 +509,10 @@ class ExplicitQuadrature(Quadrature):
             )
 
         self._points_per_cell = points.shape[1]
+
         self._whole_geo = points.shape[0] == domain.geometry_element_count()
+
+        super().__init__(domain, self._points_per_cell)
         self._points = points
         self._weights = weights